+++ src/Tasks/Monitor/HealthChecks/MaximumStorageInMegabytes.php
@@ -29,7 +29,7 @@ class MaximumStorageInMegabytes extends HealthCheck
         );
     }
 
-    protected function exceedsAllowance(int $usageInBytes): bool
+    protected function exceedsAllowance(float $usageInBytes): bool
     {
         return $usageInBytes > $this->bytes($this->maximumSizeInMegaBytes);
     }
@@ -39,7 +39,7 @@ class MaximumStorageInMegabytes extends HealthCheck
         return $megaBytes * 1024 * 1024;
     }
 
-    protected function humanReadableSize(int $sizeInBytes): string
+    protected function humanReadableSize(float $sizeInBytes): string
     {
         return Format::humanReadableSize($sizeInBytes);
     }
